requirements.txt
alembic==1.6.5
atomicwrites==1.4.0
attrs==21.2.0
click==8.0.1
colorama==0.4.4
coverage==5.5
Flask==2.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
Flask-Seeder==1.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.6.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.2
marshmallow-sqlalchemy==0.26.1
nose2==0.10.0
packaging==21.0
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.20
toml==0.10.2
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
